requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
av==15.1.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
ctranslate2==4.6.1
cycler==0.12.1
cymem==2.0.11
Cython==3.2.1
dtw-python==1.7.2
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
faster-whisper==1.2.1
filelock==3.18.0
flatbuffers==25.9.23
fonttools==4.58.0
fsspec==2025.5.1
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
hf-xet==1.1.2
huggingface-hub==0.32.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.0
keras==3.10.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.12
namex==0.1.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
openai-whisper==20250625
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
preshed==3.0.9
protobuf==6.32.1
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.1
torch==2.8.0
tqdm==4.67.1
transformers==4.52.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
whisper-timestamped==1.15.9
wrapt==1.17.2
yt-dlp==2025.10.14
zipp==3.21.0
